Create a deprecated ae module with a downgrade function
